portfolio factored out of algomanager into separate class
